Added new feature to the bot. intern.js where is defined intern commands to be used in discord
